Pull request #4017: [BLUETOOTH-17598] APP: Fix remaining TPM MISRA issue
Merge in CONNINT/bluetooth_public from BLUETOOTH-17598-misra2 to develop/main

Squashed commit of the following:

commit f8647cc5a9c87a95ab39465771a12059d7012324
Author: Silviu Petria <silviu.petria@nxp.com>
Date:   Tue May 12 10:37:01 2026 +0300

    [BLUETOOTH-17598] APP: Fix remaining TPM MISRA issue

+++ profiles/tpm/tpm_service.c
@@ -243,7 +243,7 @@ void Tpms_UpdateAdvData(uint8_t *pData, tpmsSensorReadData_t *pSensorReadData)
         (void)GattDb_WriteAttribute((uint16_t)value_tire_pressure, (uint16_t)sizeof(pSensorReadData->tirePressure), (const uint8_t*)&pSensorReadData->tirePressure);
         FLib_MemCpy(&pData[TPMS_AD_OFFSET_PRESSURE], (uint8_t*)&pSensorReadData->tirePressure, 2U);
         FLib_MemCpy(&pData[TPMS_AD_OFFSET_PRESSURE_ACC], (uint8_t*)&pSensorReadData->tirePressureAccuracy, 1U);
-        (void)GattDb_WriteAttribute((uint16_t)value_tire_temperature, sizeof(pSensorReadData->tireTemperature), (const uint8_t*)&pSensorReadData->tireTemperature);
+        (void)GattDb_WriteAttribute((uint16_t)value_tire_temperature, (uint16_t)sizeof(pSensorReadData->tireTemperature), (const uint8_t*)&pSensorReadData->tireTemperature);
         FLib_MemCpy(&pData[TPMS_AD_OFFSET_TEMPERATURE], (uint8_t*)&pSensorReadData->tireTemperature, 1U);
         FLib_MemCpy(&pData[TPMS_AD_OFFSET_TEMP_ACC], (uint8_t*)&pSensorReadData->tireTemperatureAccuracy, 1U);
